#bugfix - Parametric crash from Paul 8/20/15 - caused when using the "Add" variable in quick setup to remove variables by unchecking. Workaround is to use the "Remove" button in the parametric quick setup to remove variables. #bugfix - debug assert errors from wxWidgets fixed by type casting size_t to int in wxString::Format statements in graph.cpp and simulation.cpp

+++ src/simulation.cpp
@@ -1103,8 +1103,8 @@ void ThreadProgressDialog::Reset()
 {
 	for( size_t i=0;i<m_progbars.size();i++ )
 	{
-		m_labels[i]->SetLabel( wxString::Format("Process %d", i+1) );
-		m_progbars[i]->SetValue( 0 );
+		m_labels[i]->SetLabel(wxString::Format("Process %d", (int)(i + 1)));
+		m_progbars[i]->SetValue(0);
 		m_percents[i]->ChangeValue( wxEmptyString );
 	}
